+++ packages/trpc/server/router.ts
@@ -4,7 +4,6 @@ import { auditLogsRouter } from "./routers/auditLogs/_route";
 import { authRouter } from "./routers/auth/_route";
 import { configRouter } from "./routers/config/_route";
 import { devicesRouter } from "./routers/devices/_route";
-import { faceIdRouter } from "./routers/faceId/_route";
 import { globalReportsRouter } from "./routers/globalReports/_route";
 import { inventoryRouter } from "./routers/inventory/_route";
 import { oneTimeLoginCodesRouter } from "./routers/oneTimeLoginCodes/_route";
@@ -50,7 +49,6 @@ export const appRouter = router({
 	globalReports: globalReportsRouter,
 	tickets: ticketsRouter,
 	security: securityRouter,
-	faceId: faceIdRouter,
 });
 
 export type AppRouter = typeof appRouter;
